bug fix inside fileLoader: upload yaml file from windows
