+++ src/components/scopedVariables/ScopedVariablesLoader.tsx
@@ -10,7 +10,7 @@ export default function ScopedVariablesLoader({ status, progress, fileData, abor
     return (
         <div className="cursor-default w-200 flex column dc__align-start dc__content-start dc__gap-8">
             <div className="flex dc__align-start dc__gap-8 dc__align-self-stretch dc__content-start">
-                <p className="dc__ellipsis-right dc__no-shrin flex-grow-1 m-0 cn-6 fs-13 fw-4 lh-20 dc__align-left">
+                <p className="dc__ellipsis-right dc__no-shrink flex-grow-1 m-0 cn-6 fs-13 fw-4 lh-20 dc__align-left">
                     {fileData?.name}
                 </p>
